Improve Login and Register page
