+++ frontend/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@hookform/resolvers": "^3.1.0",
     "@iconscout/react-unicons": "^1.1.6",
     "@mantine/core": "^4.2.12",
     "@mantine/hooks": "^4.2.12",
@@ -15,7 +16,8 @@
     "react-hook-form": "^7.43.9",
     "react-router-dom": "6",
     "react-scripts": "5.0.0",
-    "web-vitals": "^2.1.0"
+    "web-vitals": "^2.1.0",
+    "yup": "^1.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
